+++ GearGauge/Models/Tags.cs
@@ -19,3 +19,4 @@ namespace GearGauge.Models
         }
     }
 }
+
