+++ GearGauge/Views/GearInventory/Add.cshtml
@@ -1,8 +1,8 @@
 ﻿@model GearGauge.ViewModels.AddGearInventoryViewModel
 
-<h1>Add Gear Inventory</h1>
+<h1>Add New Gear</h1>
 
-<form asp-action="Add" method="post" enctype="multipart/form-data">
+<form asp-controller="GearInventory" asp-action="Add" method="post" enctype="multipart/form-data">
     <div class="form-group">
         <label asp-for="Title"></label>
         <input asp-for="Title" class="form-control" />
@@ -19,19 +19,23 @@
         <span asp-validation-for="MarketValue" class="text-danger"></span>
     </div>
     <div class="form-group">
-        <label asp-for="ImageFile"></label>
-        <input asp-for="ImageFile" type="file" class="form-control" />
+        <label asp-for="ImageFile">Upload Image</label>
+        <input asp-for="ImageFile" class="form-control-file" type="file" />
         <span asp-validation-for="ImageFile" class="text-danger"></span>
     </div>
     <div class="form-group">
-        <label asp-for="SelectedTagIds"></label>
-        <select asp-for="SelectedTagIds" class="form-control" asp-items="@(new MultiSelectList(Model.AvailableTags, "Value", "Text"))" multiple="multiple"></select>
-        <span asp-validation-for="SelectedTagIds" class="text-danger"></span>
+        <label asp-for="SelectedTagIds">Select Tags</label>
+        <select asp-for="SelectedTagIds" class="form-control" asp-items="Model.AvailableTags" multiple></select>
     </div>
     <button type="submit" class="btn btn-primary">Add Gear</button>
 </form>
 
-@section Scripts {
-    <partial name="_ValidationScriptsPartial" />
+@if (Model.ImagePath != null)
+{
+    <h3>Image Preview</h3>
+    <img src="@Model.ImagePath" alt="@Model.Title" style="width: 200px; height: auto;" />
 }
 
+@section Scripts {
+    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
+}
