+++ GearGauge/Views/GearInventory/Index.cshtml
@@ -1,43 +1,35 @@
-﻿@using GearGauge.Models
-@model List<GearInventory>
+﻿@model List<GearGauge.Models.GearInventory>
 
-    <h1>Gear Gauge</h1>
+<h1>Gear Inventory</h1>
 
-    <p>
-        <a asp-controller="GearInventory" asp-action="Add">Add Gear Inventory</a>
-    </p>
+<p>
+    <a asp-controller="GearInventory" asp-action="Add">Add New Gear</a>
+</p>
 
-    @if(Model.Count == 0)
-    {
-        <p>No Gear Inventory yet!</p>
-    }
-    else
-    {
-        <table class="table">
+<table class="table">
+    <thead>
+        <tr>
+            <th>Image</th>
+            <th>Title</th>
+            <th>Description</th>
+            <th>Market Value</th>
+        </tr>
+    </thead>
+    <tbody>
+        @foreach (var item in Model)
+        {
             <tr>
-                <th>
-                    Title
-                </th>
-                <th>
-                    Description
-                </th>
-                <th>
-                    MarketValue
-                </th>
-                <th>
-                    Image
-                </th>
+                <td>
+                    <img src="@Url.Content(item.ImagePath)" alt="@item.Title" style="width: 100px; height: auto;" />
+                </td>
+                <td>@item.Title</td>
+                <td>@item.Description</td>
+                <td>@item.MarketValue</td>
+                <td>
+                    <a asp-controller="GearInventory" asp-action="Detail" asp-route-id="@item.Id">View</a> |
+                    <a asp-controller="GearInventory" asp-action="Delete" asp-route-id="@item.Id">Delete</a>
+                </td>
             </tr>
-            @foreach (var evt in Model)
-            {
-                <tr>
-                   
-                    <td><a asp-controller="GearInventory" asp-action="Detail" asp-route-id="@evt.Id">@evt.Title</a></td>
-                    <td>@evt.Description</td>
-                    <td>@evt.MarketValue</td>
-                    <td>@evt.ImageFile</td>
-                </tr>
-            }
-        </table>
-
-    }
+        }
+    </tbody>
+</table>
